Protect split of tasks from being split by 1. Just pass on the existing task object if split by 1. Also further Mandelbrot integration WIP

+++ CFMandelbrot/source/MandelbrotResult.hpp
@@ -8,7 +8,7 @@ public:
 	MandelbrotResult() {};
 	~MandelbrotResult() {};
 
-	std::vector<sf::Uint32> numbers;
+	std::vector<sf::Uint8> numbers;
 
 	inline std::string getSubtype() const { return "MandelbrotResult"; };
